Move handlers and teamsMock inside team folder
