+++ src/styles/styles.css
@@ -1,22 +1,13 @@
-* ::after,
+*,
+::after,
 ::before {
   box-sizing: border-box;
 }
 
 body {
   margin: 0;
-  font-family:
-    Ubuntu,
-    system-ui,
-    -apple-system,
-    BlinkMacSystemFont,
-    "Segoe UI",
-    Roboto,
-    Oxygen,
-    Cantarell,
-    "Open Sans",
-    "Helvetica Neue",
-    sans-serif;
+  font-family: Ubuntu, system-ui, BlinkMacSystemFont, "Segoe UI", Roboto,
+    Oxygen, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
 }
 
 h1,
